Refactor version 1.5 into the single one
