+++ lib/regaliator/account.rb
@@ -1,0 +1,9 @@
+require 'regaliator/endpoint'
+
+module Regaliator
+  class Account < Endpoint
+    def info
+      request('/account').post
+    end
+  end
+end
