--- lib/regaliator/v15.rb
@@ -1,7 +1,0 @@
-require 'regaliator/v15/client'
-
-module Regaliator
-  module V15
-    API_VERSION = '1.5'.freeze
-  end
-end
