--- lib/regaliator/v15/account.rb
@@ -1,11 +1,0 @@
-require 'regaliator/endpoint'
-
-module Regaliator
-  module V15
-    class Account < Endpoint
-      def info
-        request('/account').post
-      end
-    end
-  end
-end
