--- lib/regaliator/v15/bill.rb
@@ -1,23 +1,0 @@
-require 'regaliator/endpoint'
-
-module Regaliator
-  module V15
-    class Bill < Endpoint
-      def index(params = {})
-        request('/bills', params).post
-      end
-
-      def consult(params = {})
-        request('/bill/consult', params).post
-      end
-
-      def pay(params = {})
-        request('/bill/pay', params).post
-      end
-
-      def check(params = {})
-        request('/bill/check', params).post
-      end
-    end
-  end
-end
