--- lib/regaliator/v15/biller.rb
@@ -1,15 +1,0 @@
-require 'regaliator/endpoint'
-
-module Regaliator
-  module V15
-    class Biller < Endpoint
-      def topups(params = {})
-        request('/billers/topups', params).post
-      end
-
-      def utilities(params = {})
-        request('/billers/utilities', params).post
-      end
-    end
-  end
-end
