--- lib/regaliator/v15/rate.rb
@@ -1,15 +1,0 @@
-require 'regaliator/endpoint'
-
-module Regaliator
-  module V15
-    class Rate < Endpoint
-      def list
-        request('/rates').post
-      end
-
-      def history
-        request('/rates/history').post
-      end
-    end
-  end
-end
