+++ test/fixtures/vcr_cassettes/rate/history.yml
