+++ test/fixtures/vcr_cassettes/rate/list.yml
